Paytable name displayed in title after creation.

+++ app/src/main/java/edu/cnm/deepdive/videopoker/model/dao/PokerHandDao.java
@@ -17,6 +17,6 @@ public interface PokerHandDao {
   List<PokerHand> selectPokerHandsByBetOne();
 
   @Query("SELECT * FROM PokerHand WHERE paytable_id IS :paytableId ORDER BY betOneValue DESC")
-  List<PokerHand> selectPokerHandsByBetOneFromPaytable(int paytableId);
+  List<PokerHand> selectPokerHandsByBetOneFromPaytable(long paytableId);
 
 }
